server/requirements.txt
fastapi[standard]
email-validator
python-keycloak
loguru
sqlmodel
captcha
